add ClassifierContainer Move the store setup code into `ClassifierContainer`. `Classifier` renders after the store has loaded, and reacts to prop changes from the parent app.
